feat: create oop tasks.py
